+++ R/expect_ggplot.R
@@ -1,0 +1,8 @@
+# helper function for tests
+
+expect_ggplot <- function(g) {
+  requireNamespace("testthat")
+  testthat::expect_is(g, "ggplot")
+
+  ggsave(nullfile(), g, device = "pdf")
+}