+++ R/milestone_palette.R
@@ -6,25 +6,28 @@ sort_by_hue <- function(hex_colors) {
 
 
 
-#' @importFrom RColorBrewer brewer.pal
 #' @importFrom grDevices rainbow rgb2hsv col2rgb
-#' @importFrom rje cubeHelix
 milestone_palette_list <- list(
-  cubeHelix = function(n) rje::cubeHelix(n = n),
+  cubeHelix = function(n) {
+    requireNamespace("rje")
+    rje::cubeHelix(n = n)
+  },
   Set3 = function(n) {
+    requireNamespace("RColorBrewer")
     cols <- RColorBrewer::brewer.pal(max(3, n), "Set3")[seq_len(n)]
     sort_by_hue(cols)
   },
-  rainbow = function(n) grDevices::rainbow(n = n),
+  rainbow = function(n) {
+    grDevices::rainbow(n = n)
+  },
   auto = function(n) {
     if (n <= 12) {
       milestone_palette_list$Set3(n)
     } else {
-      # milestone_palette_list$cubeHelix(n)
-      all_colors <- grDevices::colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
+      all_colors <- grDevices::colors()[grep('gr(a|e)y', grDevices::colors(), invert = TRUE)]
       all_colors <- sort_by_hue(all_colors)[-c(1:2)] # sort and remove white/black
       ix <- ceiling(seq(0, length(all_colors), length(all_colors)/(n+1)))
-      all_colors[head(ix, -1)]
+      all_colors[utils::head(ix, -1)]
     }
   }
 )
@@ -41,7 +44,12 @@ milestone_palette <- function(name, n) {
 #'
 #' @keywords plot_helpers
 #'
+#' @returns The names of supported palettes.
+#'
 #' @export
+#'
+#' @examples
+#' get_milestone_palette_names()
 get_milestone_palette_names <- function() {
   names(milestone_palette_list)
 }