+++ R/plot_graph.R
@@ -4,9 +4,9 @@
 #' @inheritParams add_cell_coloring
 #' @inheritParams add_milestone_coloring
 #' @inheritParams plot_dimred
-#' @param transition_size The size of the transition lines between milestones.
-#' @param milestone_size The size of milestones.
-#' @param arrow_length length of the arrow.
+#' @param size_trajectory The size of the transition lines between milestones.
+#' @param size_milestones The size of milestones.
+#' @param arrow The type and size of arrow in case of directed trajectories. Set to NULL to remove arrow altogether.
 #' @param plot_milestones Whether to plot the milestones.
 #' @param alpha_cells The alpha of the cells.
 #' @param size_cells The size of the cells.
@@ -23,6 +23,8 @@
 #' @include add_coloring.R
 #'
 #' @export
+#'
+#' @returns A graph ggplot of a trajectory.
 #'
 #' @examples
 #' data(example_disconnected)
@@ -49,18 +51,18 @@ plot_graph <- dynutils::inherit_default_params(
     expression_source,
     milestones,
     milestone_percentages,
-    transition_size = 3,
-    milestone_size = 5,
+    size_trajectory = 3,
+    size_milestones = 8,
     alpha_cells = 1,
     size_cells = 2.5,
     border_radius_percentage = .1,
-    arrow_length = grid::unit(1, "cm"),
+    arrow = grid::arrow(length = grid::unit(1, "cm"), type = "closed"),
     label_milestones = dynwrap::is_wrapper_with_milestone_labelling(trajectory),
     plot_milestones = FALSE,
     adjust_weights = FALSE
   ) {
     # make sure a trajectory was provided
-    testthat::expect_true(dynwrap::is_wrapper_with_trajectory(trajectory))
+    assert_that(dynwrap::is_wrapper_with_trajectory(trajectory))
 
     # TODO: 'milestones', in this function, is both used as the colouring of the cells (which could be from a different trajectory),
     # and plotting the milestones in the same dimred as the cells.
@@ -100,7 +102,7 @@ plot_graph <- dynutils::inherit_default_params(
     if (cell_coloring_output$color_cells == "milestone") {
       milestone_positions <- left_join(milestone_positions, milestones, "milestone_id")
     } else {
-      milestone_positions$color <- NA
+      milestone_positions$color <- NA_character_
     }
 
     # get information of segments
@@ -113,14 +115,14 @@ plot_graph <- dynutils::inherit_default_params(
 
       # Divergence gray backgrounds
       geom_polygon(
-        aes(x = comp_1, y = comp_2, group = triangle_id),
+        aes(x = .data$comp_1, y = .data$comp_2, group = .data$triangle_id),
         dimred_traj$divergence_polygon_positions,
         fill = "#eeeeee"
       ) +
 
       # Divergence dashed lines
       geom_segment(
-        aes(x = comp_1_from, xend = comp_1_to, y = comp_2_from, yend = comp_2_to),
+        aes(x = .data$comp_1_from, xend = .data$comp_1_to, y = .data$comp_2_from, yend = .data$comp_2_to),
         dimred_traj$divergence_edge_positions,
         colour = "darkgray",
         linetype = "dashed"
@@ -128,58 +130,81 @@ plot_graph <- dynutils::inherit_default_params(
 
     if (plot_milestones) {
       plot <- plot +
-        geom_point(aes(comp_1, comp_2), size = 12, data = milestone_positions, colour = "gray")
+        geom_point(
+          aes(.data$comp_1, .data$comp_2),
+          size = 12,
+          data = milestone_positions,
+          colour = "gray"
+        )
     }
+
+    # add arrow if directed
+    my_arrow <-
+      if (any(trajectory$milestone_network$directed)) {
+        arrow
+      } else {
+        NULL
+      }
 
       # Transition gray border
     plot <- plot +
       geom_segment(
-        aes(x = comp_1_from, xend = comp_1_to, y = comp_2_from, yend = comp_2_to),
+        aes(x = .data$comp_1_from, xend = .data$comp_1_to, y = .data$comp_2_from, yend = .data$comp_2_to),
         dimred_segments,
-        size = transition_size + 2,
+        size = size_trajectory + 2,
         colour = "grey"
       ) +
 
       # Transition halfway arrow
       geom_segment(
-        aes(x = comp_1_from, xend = comp_1_from + (comp_1_to - comp_1_from) / 1.5, y = comp_2_from, yend = comp_2_from + (comp_2_to - comp_2_from) / 1.5),
-        dimred_segments %>% filter(directed, length > 0),
+        aes(
+          x = .data$comp_1_from,
+          xend = .data$comp_1_from + (.data$comp_1_to - .data$comp_1_from) / 1.5,
+          y = .data$comp_2_from,
+          yend = .data$comp_2_from + (.data$comp_2_to - .data$comp_2_from) / 1.5
+        ),
+        dimred_segments %>% filter(.data$directed, .data$length > 0),
         size = 1,
         colour = "grey",
-        arrow = arrow(length = arrow_length, type = "closed")
+        arrow = my_arrow
       ) +
 
       # Transition white tube
       geom_segment(
-        aes(x = comp_1_from, xend = comp_1_to, y = comp_2_from, yend = comp_2_to),
+        aes(x = .data$comp_1_from, xend = .data$comp_1_to, y = .data$comp_2_from, yend = .data$comp_2_to),
         dimred_segments,
-        size = transition_size,
+        size = size_trajectory,
         colour = "white"
       )
 
     if (plot_milestones) {
       plot <- plot +
         # Milestone white bowl
-        geom_point(aes(comp_1, comp_2), size = 10, data = milestone_positions, colour = "white") +
+        geom_point(aes(.data$comp_1, .data$comp_2), size = size_milestones, data = milestone_positions, colour = "white") +
 
         # Milestone fill
-        geom_point(aes(comp_1, comp_2, colour = color), size = 8, data = milestone_positions %>% filter(!is.na(color)), alpha = .5)
+        geom_point(
+          aes(.data$comp_1, .data$comp_2, colour = .data$color),
+          size = size_milestones * .8,
+          data = milestone_positions %>% filter(!is.na(.data$color)),
+          alpha = .5
+        )
     }
 
     # plot the cells
     if (border_radius_percentage > 0) {
       plot <- plot +
-        geom_point(aes(comp_1, comp_2), size = size_cells, color = "black", data = cell_positions)
+        geom_point(aes(.data$comp_1, .data$comp_2), size = size_cells, color = "black", data = cell_positions)
     }
 
     if (alpha_cells < 1) {
       plot <- plot +
-        geom_point(aes(comp_1, comp_2), size = size_cells * (1 - border_radius_percentage), color = "white", data = cell_positions)
+        geom_point(aes(.data$comp_1, .data$comp_2), size = size_cells * (1 - border_radius_percentage), color = "white", data = cell_positions)
     }
 
     plot <- plot +
       # Cell fills
-      geom_point(aes(comp_1, comp_2, color = color), size = size_cells * (1 - border_radius_percentage), alpha = alpha_cells, data = cell_positions) +
+      geom_point(aes(.data$comp_1, .data$comp_2, color = .data$color), size = size_cells * (1 - border_radius_percentage), alpha = alpha_cells, data = cell_positions) +
 
       color_scale +
       theme_graph() +
@@ -190,15 +215,12 @@ plot_graph <- dynutils::inherit_default_params(
 
     if (length(label_milestones)) {
       milestone_labels <- milestone_positions %>%
-        mutate(label = label_milestones[milestone_id]) %>%
-        filter(!is.na(label))
+        mutate(label = label_milestones[.data$milestone_id]) %>%
+        filter(!is.na(.data$label))
 
-      plot <- plot + geom_label(aes(comp_1, comp_2, label = label), data = milestone_labels)
+      plot <- plot + geom_label(aes(.data$comp_1, .data$comp_2, label = .data$label), data = milestone_labels)
     }
 
     plot
   }
 )
-
-#' @export
-plot_default <- plot_graph