+++ tests/testthat.R
@@ -1,8 +1,4 @@
 library(testthat)
 library(dynplot)
-library(dplyr)
-library(ggplot2)
-library(tibble)
-library(readr)
 
 test_check("dynplot")